Added basic validation method to fetcher service
